// src/types/entities.ts
export type TodolistType = {
    id: string
    title: string
    addedDate: string
    order: number
    tasks: Array<TaskType>
}

export type TaskType = {
    description: string
    title: string
    completed: boolean
    status: number
    priority: number
    startDate: string
    deadline: string
    id: string
    todoListId: string
    order: number
    addedDate: string
}

export type TodolistUpdateObject = {
    status?: number
    title?: string
    id?: string
    todoListId?: string
}